pyproject.toml
[tool.poetry]
name = "strike"
version = "0.1.0"
description = ""
authors = ["Aleksei Khatkevich <qwerty12345@disroot.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "> 0.95 < 1.0"
uvicorn = "0.22.*"
uvloop = "0.17.*"
httptools = "0.6.*"
sqlalchemy = "2.0.*"
orjson = "3.*"
asyncpg = "0.28.*"
alembic = "^1.11"
python-dotenv = "^1.0.0"
email-validator = "^2.0.0"
bcrypt = "4.*"
pyjwt = {extras = ["crypto"], version = "2.*"}
yarl = "1.9.*"
aiopath = "0.6.*"
aioshutil = "^1.3"
pyyaml = "6.*"
loguru = "0.7.*"
python-multipart = "^0.0.6"
slowapi = "0.1.*"
redis = {extras = ["hiredis"], version = "^4.5.5"}
geoalchemy2 = "^0.14"
shapely = "^2.0.1"
pyproj = "3.6.*"
fastapi-pagination = "0.12.*"
setuptools = "67.0.0"  # косяк с 68+, ждать
pydantic-settings = "^2.0.1"


[tool.poetry.group.dev.dependencies]
pytest-env = "*"
pytest = "7.3.*"
pytest-randomly = "*"
httpx = "0.24.*"
pytest-asyncio = "*"
factory-boy = "*"
pytest-factoryboy = "*"
bump-pydantic = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
